requirements.txt
alembic==1.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
enum34==1.1.6
exif==0.8.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.19
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-telegram-bot==12.2.0
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
tornado==6.0.3
urllib3==1.25.6
Werkzeug==0.16.0
yadisk==1.2.14
